Rename TestSpot to ComponentTests
This also removes the redundant hybrid tests
